requirements.txt
discord.py==1.6.0
sentry-sdk==1.0.0
tortoise-orm==0.17.0
asyncpg==0.22.0
python-dotenv==0.15.0
aerich==0.5.1
